Add generic type to Observable target.
This type allows the user to correctly type the events that are emitted by the observable. This does not add type safety since the serialized data is coerce from the any type.
TODO: fully type the log pipeline so errors will be correctly emitted at compile time if you try to log the wrong kind of data.